add onclick listener to the address textview that launches the map, move the function to the corresponding activity
