+++ app/src/main/res/layout/activity_info.xml
@@ -24,9 +24,11 @@
         android:layout_width="0dp"
         android:layout_height="30dp"
         android:layout_marginLeft="16dp"
-        android:layout_marginRight="8dp"
         android:layout_marginTop="16dp"
+        android:layout_marginRight="8dp"
+        android:clickable="true"
         android:gravity="center_vertical"
+        android:onClick="createMapIntent"
         android:text="618 E South St Orlando, FL"
         android:textColor="?android:attr/actionMenuTextColor"
         android:textSize="18sp"
